+++ libdeepin-font-installer/dfontmanager.h
@@ -44,7 +44,7 @@ public:
 
 signals:
     void output(const QString &output);
-    void installing(const QString &filePath, const float &percent);
+    void installing(const QString &filePath, const double &percent);
     void singleInstallFinished();
     void installFinished();
     void installChanged(const QString &instPath);
